Push inputs to URL string
REfactored back name and birthday to be part of a reducer. Still need
to move the builder function into the impl as a `new` method.

This is currently broken, as the page reloads every time an input
changes. Also, it is not setting the initial state from the URL query.